+++ packages/mobile/src/store/store.keys.ts
@@ -3,4 +3,5 @@ export enum StoreKeys {
   Navigation = 'Navigation',
   NativeServices = 'NativeServices',
   PushNotifications = 'PushNotifications',
+  Keys = 'Keys',
 }