+++ packages/types/src/keys.ts
@@ -1,0 +1,10 @@
+import { Base58, KeyMetadata } from '@localfirst/crdx'
+
+export interface StorableKey {
+  keyName: string
+  key: string | Base58
+}
+
+export interface KeysUpdatedEvent {
+  keys: StorableKey[]
+}